+++ modules/portal/portal-sso-openid/src/com/liferay/portal/sso/openid/internal/servlet/taglib/OpenIdNavigationPreDynamicInclude.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.sso.openid.servlet.taglib;
+package com.liferay.portal.sso.openid.internal.servlet.taglib;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
